requirements.txt
Django==2.2.6
django-datatables-view==1.19.1
django-widget-tweaks==1.4.8
pytz==2020.1
sqlparse==0.3.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.5.0
django-extensions==3.0.9
drf-yasg==1.17.1
django-filter==2.1.0
Unidecode==1.1.2
PyJWT==1.7.1
pandas==1.2.3
numpy==1.20.1
xlrd==1.2.0
